Added abstraction to handle code duplication
