+++ lib/rest_endpoint/swagger2/rest_swagger_final_path_processing.py
@@ -2,7 +2,7 @@ import os
 import re
 import collections
 from lib import utils
-from lib.endpoint_processing import pathProcessing
+from lib.path_processing import pathProcessing
 
 class restSwaggerPathProcessing(pathProcessing):
     def __init__(self):
